Refactor cairo1-run crate (#1601)
* Create a cairo_run for cairo1-run crate + Refactor cairo1 execution into different functions

* Clippy

* Merge imports + Impl Default for Cairo1RunConfig

* Remove allow unused imports

* Add changelog entry

* Use FuncArg slice instead of FuncArgs wrapper in public api

* Update changelog
